test: share external database cleanup helpers
