+++ apps/backend/src/lib/external-db-sync-queue.test.ts
@@ -248,6 +248,11 @@ describe("recordExternalDbSyncDeletion", () => {
       })).toBeNull();
     } finally {
       await retryTransaction(globalPrismaClient, async (tx) => {
+        await recordExternalDbSyncDeletion(tx, {
+          tableName: "EmailOutbox",
+          tenancyId,
+          emailOutboxId: id,
+        });
         await tx.emailOutbox.deleteMany({
           where: { tenancyId, id },
         });
